Add Partners to index page
