+++ src/components/PartnersModule/index.js
@@ -1,0 +1,1 @@
+export { PartnersModule as default } from './PartnersModule';
